Silence compile-time or run-time warnings.
See:

* https://duckduckgo.com/?q=warnings+programming&atb=v140-1&ia=web

They are a distraction at best and may indicate other issues.

+++ fc-solve/source/board_gen/pi_make_microsoft_freecell_board.c
@@ -32,7 +32,8 @@ int main(int argc, char *argv[])
         print_ts = true;
         ++arg;
     }
-    const long long gamenumber = ((arg < argc) ? atoll(argv[arg++]) : 1);
+    const unsigned long long gamenumber =
+        ((arg < argc) ? (unsigned long long)atoll(argv[arg++]) : 1);
 
     fcs_state_string s;
     get_board_l(gamenumber, s);
