+++ fc-solve/source/measure_depth_dep_tests_order_performance.c
@@ -167,7 +167,7 @@ static inline int range_solvers_main(int argc, char *argv[], int arg,
              board_num <= end_board; board_num++, curr_result++)
         {
             fcs_state_string state_string;
-            get_board(board_num, state_string);
+            get_board((unsigned long long)board_num, state_string);
 
 #ifndef FCS_WITHOUT_MAX_NUM_STATES
             freecell_solver_user_limit_iterations_long(instance, iters_limit);
